Fix jest dependency issue and moved integration test to App, passing
