+++ package.json
@@ -38,7 +38,6 @@
     ]
   },
   "devDependencies": {
-    "@testing-library/jest-dom": "^5.11.4",
-    "jest": "^26.9.0"
+    "@testing-library/jest-dom": "^5.11.4"
   }
 }
