+++ src/UIComponents/Header/Header.css
@@ -1,0 +1,6 @@
+header {
+  width: 100vw;
+  height: 10vh;
+  border: 1px solid blue;
+  font-size: 3vh;
+}
