+++ src/UIComponents/Header/Header.js
@@ -1,4 +1,5 @@
 import React from 'react'
+import './Header.css'
 
 const Header = () => {
   return <header>Street | ART | Walk</header>
