+++ src/pages/LoginPage/LoginPage.test.js
@@ -1,14 +1,12 @@
 import React from 'react'
-import ReactDOM from 'react-dom';
-import { render, cleanup, screen, fireEvent } from '@testing-library/react'
-import '@testing-library/jest-dom';
+import ReactDOM from 'react-dom'
+import { render, cleanup, screen } from '@testing-library/react'
+import '@testing-library/jest-dom'
 import { BrowserRouter } from 'react-router-dom'
 import LoginPage from './LoginPage'
 const setIsLoggedIn = jest.fn()
 
 describe('<LoginPage/>', () => {
-  beforeEach(() => {
-  })
   afterEach(cleanup)
 
   it('should render without crashing', () => {
@@ -37,42 +35,4 @@ describe('<LoginPage/>', () => {
     expect(loginBtn).toBeInTheDocument()
     expect(signupBtn).toBeInTheDocument()
   })
-
-  // this test goes in APP
-  it('should change page with successful login', () => {
-    const { getByText, getByPlaceholderText, getByRole, debug } = render(
-      <BrowserRouter>
-        <LoginPage setIsLoggedIn={setIsLoggedIn} />
-      </BrowserRouter>)
-
-    const title = getByText('Street | ART | Walk')
-    const usernameInput = getByPlaceholderText('Username...')
-    const passwordInput = getByPlaceholderText('Password...')
-    const loginBtn = getByRole('button', { name: 'LOG IN' })
-
-    expect(title).toBeInTheDocument()
-    expect(usernameInput).toBeInTheDocument()
-    expect(passwordInput).toBeInTheDocument()
-    expect(loginBtn).toBeInTheDocument()
-    // fireEvent.change(usernameInput, {target: {value: 'testUser'}})
-    // fireEvent.change(passwordInput, {target: {value: 'testPassword'}})
-    fireEvent.click(loginBtn) 
-    debug()
-    // expect(usernameInput).not.toBeInTheDocument()
-    // expect(passwordInput).not.toBeInTheDocument()
-    // expect(loginBtn).not.toBeInTheDocument()
-    
-  })
-
-  // it('Should change locations when the log in button is clicked', async () => {
-  //   const testHistoryObject = createMemoryHistory()
-  //   const { getByRole } = render( 
-  //     <Router history={ testHistoryObject }>
-  //       <App />
-  //     </Router> )
-  //   expect(testHistoryObject.location.pathname).toEqual('/')
-  //   const logInButton = await waitFor(() => getByRole('button', {name: 'LOG IN'}))
-  //   fireEvent.click(logInButton) 
-  //   expect(testHistoryObject.location.pathname).toEqual('/login')
-  // })
 })
